+++ CHANGES.md
@@ -1,6 +1,12 @@
 # Release notes
 All notable changes to this project will be documented in this file.
 
+## 0.11.0
+- FIX: Sell cards crash 
+- Update load portfolio improve load time 
+- First land implementation
+- Add last investment marker for visual effect.
+
 ## 0.10.0
 - Add monitor of staked DEC
 - Update ordering of portfolio ALL graph